+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 
 ### Other changes
 
+* added defaultTemplate setting to server.json
 * added autodetection of piece type and size based on image dimensions in upload modal
 * fixed no grid visible in low quality mode
 * fixed empty goblin label in RPG template